Renaming src -> sentiment_model_trainer to prevent issues with src conflict with mutation testing
