+++ pytest.ini
@@ -1,5 +1,6 @@
 # pytest.ini
 [pytest]
-addopts = --cov=src --cov-report=term-missing
+addopts = --cov=sentiment_model_trainer --cov-report=term-missing
 testpaths = tests
 python_files = test_*.py
+pythonpath = sentiment_model_trainer
