+++ run_all_test.bash
@@ -1,12 +1,16 @@
 echo "=== Running pytest with coverage ==="
-pytest --cov=src --cov-report=term-missing > test_report.txt
+pytest --cov=sentiment_model_trainer --cov-report=term-missing | tee test_report.txt
 COV=$(coverage report | tail -n 1 | awk '{print $4}')
 echo ">>> Coverage: ${COV}"
 
 echo "=== Running mutation testing (mutmut) ==="
-mutmut run src
-MUT_SCORE=$(mutmut results | grep ^Score | awk '{print $2}')
-echo ">>> Mutation Testing Score: ${MUT_SCORE}"
+mutmut run  | tee mutmut_report.txt
+
+PASSED=$(cat mutmut_report.txt | tail -n 2 | head -n 1 | grep -o "🎉 [0-9]*"| tail -n 1 | awk '{print $2}' )
+TOTAL=$(cat mutmut_report.txt | tail -n 2 | head -n 1 | grep -o "0/[0-9]*" | cut -c 3-)
+MUT_SCORE=$(awk -v p="$PASSED" -v t="$TOTAL" 'BEGIN { printf "%.2f", (p/t)*100 }')
+
+echo ">>> Mutation Passed: ${MUT_SCORE}%"
 
 BENCH_LINE=$(grep 'test_inference_latency' test_report.txt)
 MIN_LAT=$(echo $BENCH_LINE | awk '{print $2}' | tr -d ',')
@@ -33,4 +37,6 @@ cat <<EOF > metrics/test_metrics.json
 }
 EOF
 
+# Clean up temporary files
 rm test_report.txt
+rm mutmut_report.txt
