+++ setup.cfg
@@ -1,0 +1,3 @@
+[mutmut]
+paths_to_mutate=sentiment_model_trainer/
+tests_dir=tests/
